cssa/requirements.txt
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
torchtext==0.14.1
torchdata==0.5.1
torchinfo==1.8.0
numpy
pandas
Pillow
opencv-python
xmltodict
matplotlib
tqdm
